fix some old boost python issues
remove unneeded dllexports to remove warnings

+++ src/boost_python/object/pickle_support.cpp
@@ -44,7 +44,7 @@ namespace {
       result.append(initargs);
       object getstate = getattr(instance_obj, "__getstate__", none);
       object instance_dict = getattr(instance_obj, "__dict__", none);
-      long len_instance_dict = 0;
+      Py_ssize_t len_instance_dict = 0;
       if (!instance_dict.is_none()) {
           len_instance_dict = len(instance_dict);
       }
